Refactored test implicits
