--- src/test/scala/play/api/cache/redis/configuration/Implicits.scala
@@ -1,10 +1,0 @@
-package play.api.cache.redis.configuration
-
-import scala.language.implicitConversions
-
-import play.api.cache.redis.TestImplicits
-
-/**
-  * @author Karel Cemus
-  */
-object Implicits extends TestImplicits
