+++ src/test/scala/play/api/cache/redis/configuration/RedisHostSpecs.scala
@@ -6,7 +6,7 @@ import org.specs2.mutable.Spec
   * @author Karel Cemus
   */
 class RedisHostSpecs extends Spec {
-  import Implicits._
+  import play.api.cache.redis.Implicits._
 
   private implicit val loader = RedisHost
