+++ src/test/scala/play/api/cache/redis/configuration/RedisInstanceProviderSpecs.scala
@@ -6,7 +6,7 @@ import org.specs2.mutable.Specification
   * @author Karel Cemus
   */
 class RedisInstanceProviderSpecs extends Specification {
-  import Implicits._
+  import play.api.cache.redis.Implicits._
 
   val defaultCache = RedisStandalone( defaultCacheName, RedisHost( localhost, defaultPort, database = 0 ), defaults )
